Update to use DECOMPRESS/ONLY
The previous implementation would fabricate a compression envelope
which it claimed to be a ZLIB format, but had a CRC32 checked like
GZIP.  In reality it was raw deflate data and had nothing to do with gzip
or zlib envelope formats.

This uses a new /ONLY setting in Ren/C which finesses the issue, so
that the data can be passed in raw.  The size and CRC are then
checked separately after the decompression is done.